requirements.txt
coverage==4.5.2
nose==1.3.7
pip==19.0.1
py-trello==0.14.0
testfixtures==6.5.2